Refactor reportable to not take children
